refactor: move Validation to protocols folder
